Add message reaction functionality in chat component
This commit includes changes mainly in the chat component to handle emojis differently depending on whether a message is active or not. It introduces a new event handler, 'handleReactionMessage' which triggers emoji picker and sets active message. Handling of a click event for emojis is also updated to associate it with specific message as a reaction.